split files, support more protocols, fix blocking call
